requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
absl-py=0.11.0=pypi_0
addict=2.4.0=pypi_0
albumentations=0.5.2=pypi_0
appdirs=1.4.4=pypi_0
argon2-cffi=20.1.0=pypi_0
astor=0.8.1=pypi_0
astroid=2.4.2=py37_0
astunparse=1.6.3=pypi_0
async-generator=1.10=pypi_0
attrs=20.3.0=pypi_0
babel=2.9.0=pypi_0
backcall=0.2.0=pypi_0
bce-python-sdk=0.8.53=pypi_0
blas=1.0=mkl
bleach=3.2.1=pypi_0
bzip2=1.0.8=h7b6447c_0
ca-certificates=2021.1.19=h06a4308_1
cachetools=4.1.1=pypi_0
cairo=1.14.12=h8948797_3
certifi=2020.12.5=py37h06a4308_0
cffi=1.14.4=pypi_0
cfgv=3.2.0=pypi_0
chardet=3.0.4=pypi_0
click=7.1.2=pypi_0
cloudpickle=1.6.0=py_0
cma=3.0.3=pypi_0
colorama=0.4.4=pypi_0
colorlog=4.6.2=pypi_0
cv=1.0.0=pypi_0
cycler=0.10.0=pypi_0
cython=0.29.21=pypi_0
cytoolz=0.11.0=py37h7b6447c_0
dask-core=2021.3.0=pyhd3eb1b0_0
decorator=4.4.2=pyhd3eb1b0_0
defusedxml=0.6.0=pypi_0
distlib=0.3.1=pypi_0
dominate=2.6.0=pypi_0
entrypoints=0.3=pypi_0
ffmpeg=4.0=hcdf2ecd_0
filelock=3.0.12=pypi_0
flake8=3.8.4=pypi_0
flask=1.1.2=pypi_0
flask-babel=2.0.0=pypi_0
flask-cors=3.0.10=pypi_0
fontconfig=2.13.0=h9420a91_0
freeglut=3.0.0=hf484d3e_5
freetype=2.10.4=h5ab3b9f_0
funcsigs=1.0.2=pypi_0
future=0.18.2=pypi_0
gast=0.3.3=pypi_0
glib=2.63.1=h5a9c865_0
google-auth=1.23.0=pypi_0
google-auth-oauthlib=0.4.2=pypi_0
google-pasta=0.2.0=pypi_0
graphite2=1.3.14=h23475e2_0
grpcio=1.34.0=pypi_0
gunicorn=20.0.4=pypi_0
h5py=2.10.0=pypi_0
harfbuzz=1.8.8=hffaf4a1_0
hdf5=1.10.2=hba1933b_1
icu=58.2=he6710b0_3
identify=1.5.12=pypi_0
idna=2.10=pypi_0
imageio=2.9.0=py_0
imgaug=0.4.0=pypi_0
importlib-metadata=3.1.1=pypi_0
intel-openmp=2020.2=254
ipykernel=5.4.2=pypi_0
ipython=7.19.0=pypi_0
ipython-genutils=0.2.0=pypi_0
ipywidgets=7.5.1=pypi_0
isort=5.6.4=py_0
itk=5.1.2=pypi_0
itk-core=5.1.2=pypi_0
itk-filtering=5.1.2=pypi_0
itk-io=5.1.2=pypi_0
itk-numerics=5.1.2=pypi_0
itk-registration=5.1.2=pypi_0
itk-segmentation=5.1.2=pypi_0
itsdangerous=1.1.0=pypi_0
jasper=2.0.14=h07fcdf6_1
jedi=0.17.2=pypi_0
jinja2=2.11.2=pypi_0
joblib=1.0.0=pypi_0
jpeg=9b=h024ee3a_2
jsonpatch=1.28=pypi_0
jsonpointer=2.0=pypi_0
jsonschema=3.2.0=pypi_0
jupyter=1.0.0=pypi_0
jupyter-client=6.1.7=pypi_0
jupyter-console=6.2.0=pypi_0
jupyter-core=4.7.0=pypi_0
jupyterlab-pygments=0.1.2=pypi_0
keras=2.4.3=pypi_0
keras-contrib=2.0.8=pypi_0
keras-preprocessing=1.1.2=pypi_0
kiwisolver=1.3.1=py37h2531618_0
lazy-object-proxy=1.4.3=py37h7b6447c_0
lcms2=2.11=h396b838_0
libedit=3.1.20191231=h14c3975_1
libffi=3.2.1=hf484d3e_1007
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.3.0=hdf63c60_0
libglu=9.0.0=hf484d3e_1
libopencv=3.4.2=hb342d67_1
libopus=1.3.1=h7b6447c_0
libpng=1.6.37=hbc83047_0
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.1.0=h2733197_1
libuuid=1.0.3=h1bed415_2
libvpx=1.7.0=h439df22_0
libxcb=1.14=h7b6447c_0
libxml2=2.9.10=hb55368b_3
lz4-c=1.9.2=heb0550a_3
markdown=3.3.3=pypi_0
markupsafe=1.1.1=pypi_0
matplotlib-base=3.3.4=py37h62a2d02_0
mccabe=0.6.1=py37_1
mistune=0.8.4=pypi_0
mkl=2020.2=256
mkl-service=2.3.0=py37he8ac12f_0
mkl_fft=1.2.0=py37h23d657b_0
mkl_random=1.1.1=py37h0573a6f_0
mmcv=1.2.5=pypi_0
nbclient=0.5.1=pypi_0
nbconvert=6.0.7=pypi_0
nbformat=5.0.8=pypi_0
ncurses=6.2=he6710b0_1
nest-asyncio=1.4.3=pypi_0
networkx=2.5=py_0
neurite=0.1=pypi_0
nibabel=3.2.1=pypi_0
nltk=3.5=pypi_0
nodeenv=1.5.0=pypi_0
notebook=6.1.5=pypi_0
numpy=1.18.5=pypi_0
numpy-base=1.19.2=py37hfa32c7d_0
oauthlib=3.1.0=pypi_0
objgraph=3.5.0=pypi_0
olefile=0.46=py37_0
opencv=3.4.2=py37h6fd60c2_1
opencv-python=4.5.1.48=pypi_0
opencv-python-headless=4.5.1.48=pypi_0
openssl=1.0.2u=h7b6447c_0
opt-einsum=3.3.0=pypi_0
packaging=20.8=pypi_0
paddlehub=1.8.3=pypi_0
paddlepaddle=1.8.5=pypi_0
paddleslim=1.1.1=pypi_0
paddlex=1.3.3=pypi_0
pandas=1.1.5=pypi_0
pandocfilters=1.4.3=pypi_0
parso=0.7.1=pypi_0
pathlib=1.0.1=pypi_0
pcre=8.44=he6710b0_0
pexpect=4.8.0=pypi_0
pickleshare=0.7.5=pypi_0
pillow=8.0.1=pypi_0
pip=20.3=py37h06a4308_0
pixman=0.40.0=h7b6447c_0
pre-commit=2.9.3=pypi_0
prettytable=2.0.0=pypi_0
prometheus-client=0.9.0=pypi_0
prompt-toolkit=3.0.8=pypi_0
protobuf=3.14.0=pypi_0
psutil=5.8.0=pypi_0
ptyprocess=0.6.0=pypi_0
py-opencv=3.4.2=py37hb342d67_1
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pycocotools=2.0.2=pypi_0
pycodestyle=2.6.0=pypi_0
pycparser=2.20=pypi_0
pycryptodome=3.9.9=pypi_0
pyflakes=2.2.0=pypi_0
pygments=2.7.3=pypi_0
pylint=2.6.0=py37_0
pyparsing=2.4.7=pyhd3eb1b0_0
pyrsistent=0.17.3=pypi_0
pystrum=0.1=pypi_0
python=3.7.0=h6e4f718_3
python-dateutil=2.8.1=pyhd3eb1b0_0
python-graphviz=0.16=pypi_0
pytz=2020.5=pypi_0
pywavelets=1.1.1=py37h7b6447c_2
pyyaml=5.3.1=pypi_0
pyzmq=20.0.0=pypi_0
qtconsole=5.0.1=pypi_0
qtpy=1.9.0=pypi_0
rarfile=4.0=pypi_0
readline=7.0=h7b6447c_5
regex=2020.11.13=pypi_0
requests=2.25.0=pypi_0
requests-oauthlib=1.3.0=pypi_0
rsa=4.6=pypi_0
scikit-image=0.15.0=pypi_0
scikit-learn=0.23.2=pypi_0
scipy=1.5.4=pypi_0
send2trash=1.5.0=pypi_0
sentencepiece=0.1.95=pypi_0
setuptools=50.3.2=py37h06a4308_2
shapely=1.7.1=pypi_0
simpleitk=2.0.2=pypi_0
six=1.15.0=py37h06a4308_0
sklearn=0.0=pypi_0
sqlite=3.33.0=h62c20be_0
tensorboard=2.4.0=pypi_0
tensorboard-plugin-wit=1.7.0=pypi_0
tensorboardx=2.1=pypi_0
tensorflow-estimator=2.3.0=pypi_0
tensorflow-gpu=2.3.1=pypi_0
termcolor=1.1.0=pypi_0
terminado=0.9.1=pypi_0
testpath=0.4.4=pypi_0
threadpoolctl=2.1.0=pypi_0
tifffile=2020.12.8=pypi_0
tk=8.6.10=hbc83047_0
toml=0.10.1=py_0
toolz=0.11.1=pyhd3eb1b0_0
torch=1.7.1=pypi_0
torchfile=0.1.0=pypi_0
torchsummary=1.5.1=pypi_0
torchvision=0.8.2=pypi_0
tornado=6.1=py37h27cfd23_0
tqdm=4.54.1=pypi_0
traitlets=5.0.5=pypi_0
typed-ast=1.4.1=py37h7b6447c_0
typing-extensions=3.7.4.3=pypi_0
urllib3=1.26.2=pypi_0
virtualenv=20.3.0=pypi_0
visdom=0.1.8.9=pypi_0
visualdl=2.1.0=pypi_0
voxelmorph=0.1=pypi_0
wcwidth=0.2.5=pypi_0
webencodings=0.5.1=pypi_0
websocket-client=0.57.0=pypi_0
werkzeug=1.0.1=pypi_0
wheel=0.36.0=pyhd3eb1b0_0
widgetsnbextension=3.5.1=pypi_0
wrapt=1.12.1=pypi_0
xz=5.2.5=h7b6447c_0
yaml=0.2.5=h7b6447c_0
yapf=0.26.0=pypi_0
zipp=3.4.0=pypi_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.5=h9ceee32_0
